Add files via upload
This file contains the code to autologin your Zerodha API account. The files that are taken as an input contain personal information, therefore a sample file is uploaded for the same. The code is written in python. The file takes your login information as input and then creates the request token and access token afterwards. The access token is required to gain access to the account and execute algorithms on it. You have to generate the access token daily.